requirements.txt
streamlit
pandas
pyyaml
openai
pyarrow